refactor: split device classes and exceptions out of hub.py (stage 3a)
hub.py (2,9xx lines) now holds only session orchestration, polling
tiers, and the write path (~900 lines):

- exceptions.py: the SolarEdgeException hierarchy
- devices.py: SolarEdgeInverter/MMPPTUnit/Meter/Battery/EVSE plus their
  decode helpers (decode_sunspec_*, drop/log_decoded, APC field tables,
  *_DECODED_KEYS) — devices reference the hub only through the instance
  passed at construction (TYPE_CHECKING-only import, no cycle)
- const.py: the repair-issue id helpers (pure strings)

Every previously-public name is still importable from hub.py via
explicit re-exports (__all__); no test changed. Behavior-neutral by
the same gates as stage 2: decode golden, entity golden, transaction
counts, 751 tests.
