+++ custom_components/solaredge_modbus_multi/exceptions.py
@@ -1,0 +1,75 @@
+"""Exception hierarchy for SolarEdge Modbus Multi."""
+
+from __future__ import annotations
+
+
+class SolarEdgeException(Exception):
+    """Base class for other exceptions"""
+
+    pass
+
+
+class HubInitFailed(SolarEdgeException):
+    """Raised when an error happens during init"""
+
+    pass
+
+
+class DeviceInitFailed(SolarEdgeException):
+    """Raised when a device can't be initialized"""
+
+    pass
+
+
+class DeviceIsEVSE(SolarEdgeException):
+    """Raised when an inverter device matches a EVSE model"""
+
+    pass
+
+
+class ModbusReadError(SolarEdgeException):
+    """Raised when a modbus read fails (generic)"""
+
+    pass
+
+
+class ModbusIllegalFunction(SolarEdgeException):
+    """Raised when a modbus address is invalid"""
+
+    pass
+
+
+class ModbusIllegalAddress(SolarEdgeException):
+    """Raised when a modbus address is invalid"""
+
+    pass
+
+
+class ModbusIllegalValue(SolarEdgeException):
+    """Raised when a modbus address is invalid"""
+
+    pass
+
+
+class ModbusIOError(SolarEdgeException):
+    """Raised when a modbus IO error occurs"""
+
+    pass
+
+
+class ModbusWriteError(SolarEdgeException):
+    """Raised when a modbus write fails (generic)"""
+
+    pass
+
+
+class DataUpdateFailed(SolarEdgeException):
+    """Raised when an update cycle fails"""
+
+    pass
+
+
+class DeviceInvalid(SolarEdgeException):
+    """Raised when a device is not usable or invalid"""
+
+    pass
